requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Authlib==0.15.5
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
dataclasses-json==0.5.7
dateparser==1.1.2
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
frozenlist==1.3.1
idna==3.4
marshmallow==3.18.0
marshmallow-enum==1.5.1
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
pycparser==2.21
pyparsing==3.0.9
python-binance==1.0.16
python-dateutil==2.8.2
python-twitter-v2==0.7.9
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
typing-inspect==0.8.0
tzdata==2022.5
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
websockets==10.3
yarl==1.8.1
